backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
python-multipart==0.0.6
httpx==0.25.2
openai==1.3.0
anthropic==0.7.0
tiktoken==0.5.1
pandas==2.1.3
numpy==1.26.2